+++ _fixtures/internal/dir0/pkg/main.go
