+++ _fixtures/internal/dir0/renamedpackage/main.go
