+++ _fixtures/internal/dir1/pkg/main.go
